Highlighting works! By default everything marks as a "pass." To specify boundaries, a parameter DSV file needs to be created. The DSV is essentially a file where each entry is separated by a delimiter, in this case a semicolon. In the DSV, keep everything in the following format:
	     package/path/class.name.method():inst_percent:branch_percent:cxty_percent:line_percent:method_percent:class_percent:propagate

Where "*_percent" is that field's desired minimum percentage. The "propagate" field is an optional boolean value that specifies whether this parameter should apply to only the specified package/class/method or to all children as well.

Note that this is pretty buggy at the moment (e.g. you can only specify packages). But it works for the simple cases!